chore: ci.yml 더미데이터 추가 chore(auth): CustomOauth2UserServiceTest 삭제 refactor(auth): codeStyle 수정 refactor(auth): codeStyle 수정 Merge branch 'dev' into feat/13/JJY refactor(member, auth): member도메인의 logout을 auth도메인으로 이동, 그 밖에 리뷰 반영
