+++ src/main/java/com/deal4u/fourplease/domain/auth/controller/AuthController.java
@@ -2,11 +2,18 @@ package com.deal4u.fourplease.domain.auth.controller;
 
 import com.deal4u.fourplease.domain.auth.dto.TokenPair;
 import com.deal4u.fourplease.domain.auth.service.AuthService;
+import com.deal4u.fourplease.domain.member.entity.Member;
+import com.deal4u.fourplease.domain.member.service.MemberService;
+import io.swagger.v3.oas.annotations.responses.ApiResponse;
+import io.swagger.v3.oas.annotations.responses.ApiResponses;
 import lombok.RequiredArgsConstructor;
+import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
+import org.springframework.security.core.annotation.AuthenticationPrincipal;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestHeader;
 import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.ResponseStatus;
 import org.springframework.web.bind.annotation.RestController;
 
 @RestController
@@ -22,5 +29,19 @@ public class AuthController {
         String refreshToken = refreshTokenHeader.replace("Bearer ", "");
         TokenPair tokenPair = authService.refreshAccessToken(refreshToken);
         return ResponseEntity.ok(tokenPair);
+    }
+
+    @PostMapping("/logout")
+    @ApiResponses(value = {
+            @ApiResponse(responseCode = "204", description = "로그아웃 성공"),
+            @ApiResponse(responseCode = "401", description = "인증 실패 – 토큰 누락 또는 유효하지 않음"),
+            @ApiResponse(responseCode = "403", description = "이미 무효화된 토큰")
+    })
+    @ResponseStatus(HttpStatus.NO_CONTENT)
+    public ResponseEntity<Void> logout(
+            @RequestHeader("Authorization") String authHeader,
+            @AuthenticationPrincipal Member member
+    ) {
+        return authService.logout(authHeader, member);
     }
 }
